Все работает

+++ ContentRecommender.Core/Configuration/ProviderConfig.cs
@@ -5,7 +5,7 @@ public class ProviderConfig
     public string BaseUrl { get; set; } = string.Empty;
     public string ApiKey { get; set; } = string.Empty;
     public string ApiKeyHeader { get; set; } = string.Empty;
-    public int DefaultRatingFrom { get; set; } = 5;
+    public double DefaultRatingFrom { get; set; } = 5.0;
     public int DefaultLimit { get; set; } = 40;
     public Dictionary<string, string> Urls { get; set; } = new();
     public FieldMapping FieldMapping { get; set; } = new();
